requirements.txt
Django==3.0.3
pytest-django==3.8.0
